Day 3 Assignment Submission
Q1.
Result:
Before Deletion My Array : [ 2  4  6  8 10 18 20]
<class 'numpy.ndarray'>
Enter Position from where u want to delete:3
After Deletion My Array : [ 2  4  8 10 18 20]

Q2.
Summary:
Constructors are generally used for instantiating an object. The task of constructors is to initialize(assign values) to the data members of the class when an object of class is created. In Python the __init__() method is called the constructor and is always called when an object is created.